+++ urbackupserver/server.h
@@ -28,8 +28,8 @@ public:
 
 	void operator()(void);
 
-	static IPipeThrottler *getGlobalInternetThrottler(size_t speed_bps);
-	static IPipeThrottler *getGlobalLocalThrottler(size_t speed_bps);
+	static IPipeThrottler *getGlobalInternetThrottler(int speed_bps);
+	static IPipeThrottler *getGlobalLocalThrottler(int speed_bps);
 
 	static void cleanupThrottlers(void);
